IntelThermal renamed to CPUSensors. HWMonitor: added option to show disk partitions names in sensors menu

+++ SuperIOSensors/ITEIT87x.cpp
@@ -103,7 +103,7 @@ float IT87xMonitor::readVoltage(UInt32 index)
 
 float IT87xMonitor::readTachometer(UInt32 index)
 {
-    long value;
+    UInt32 value;
     
     if (has16bitFanCounter)
     {
